// frontend/src/pages/Listing/index.tsx
import axios from "axios";
import MovieCard from "components/MovieCard";
import Pagination from "components/Pagination";
import { useEffect, useState } from "react";
import { Movie, MoviePage } from "types/movie";
import { BASE_URL } from "utils/requests";

type Props = {
  deleteMovie: Function;
  movie: Movie;
};

function Listing() {
  const [pageNumber, setPageNumber] = useState(0);

  const [page, setPage] = useState<MoviePage>({
    content: [],
    last: true,
    totalPages: 0,
    totalElements: 0,
    size: 12,
    number: 0,
    first: true,
    numberOfElements: 0,
    empty: true,
  });

  useEffect(() => {
    axios
      .get(`${BASE_URL}/movies?size=12&page=${pageNumber}&sort=title`)
      .then((response) => {
        const data = response.data as MoviePage;
        setPage(data);
      });
  }, [pageNumber]);

  function deleteMovie(id: number) {
    axios.delete(`${BASE_URL}/movies/${id}`).then(() => {
      getMovies();
    });
  }

  function getMovies() {
    axios
      .get(`${BASE_URL}/movies?size=12&page=${pageNumber}&sort=id`)
      .then((response) => {
        let data = response.data as MoviePage;

        setPage(data);
      });
  }

  const handlePageChange = (newPageNumber: number) => {
    setPageNumber(newPageNumber);
  };

  return (
    <>
      <Pagination page={page} onChange={handlePageChange} />
      <div className="container">
        <div className="row">
          {page.content.map((movie) => (
            <div key={movie.id} className="col-sm-6 col-lg-4 col-xl-3 mb-3">
              <MovieCard movie={movie} deleteMovie={deleteMovie} />
            </div>
          ))}
        </div>
      </div>
    </>
  );
}

export default Listing;
